# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mechanic Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SyncRoomCode {
  code: string;
}

export interface Vehicle {
  id: string;
  registration: string;
  make: string;
  model: string;
  createdAt: string;
}

export interface Job {
  id: string;
  vehicleRegistration: string;
  date: string;
  time: string;
  description: string;
  notes: string;
  isService: boolean;
  createdAt: string;
}

export interface SyncRoomData {
  vehicles: Vehicle[];
  jobs: Job[];
  updatedAt: string;
}

export interface SyncPayload {
  vehicles: Vehicle[];
  jobs: Job[];
}

export interface SyncUpdatedAt {
  updatedAt: string;
}