# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mechanic Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: string;
  vehicleRegistration: string;
  date: string;
  time: string;
  description: string;
  notes: string;
  isService: boolean;
  createdAt: string;
}