# lib/api-zod/src/generated/types/syncRoomData.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mechanic Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';
import type { Vehicle } from './vehicle';

export interface SyncRoomData {
  vehicles: Vehicle[];
  jobs: Job[];
  updatedAt: string;
}