# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mechanic Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: string;
  registration: string;
  make: string;
  model: string;
  createdAt: string;
}